requirement.txt
requests==2.28.2
python-dotenv==0.21.0
streamlit==1.22.0
plotly==5.13.0